+++ packages/core/index.ts
@@ -46,15 +46,17 @@ export interface BaseOptions<TShared extends Record<string, ZodType>> {
 export interface LooseOptions<TShared extends Record<string, ZodType>>
   extends BaseOptions<TShared> {
   runtimeEnvStrict?: never;
+
   /**
-   * Runtime Environment variables to use for validation - `process.env`, `import.meta.env` or similar.
-   * Unlike `runtimeEnvStrict`, this doesn't enforce that all environment variables are set.
+   * What object holds the environment variables at runtime. This is usually
+   * `process.env` or `import.meta.env`.
    */
+  // Unlike `runtimeEnvStrict`, this doesn't enforce that all environment variables are set.
   runtimeEnv: Record<string, string | boolean | number | undefined>;
 }
 
 export interface StrictOptions<
-  TPrefix extends string,
+  TPrefix extends string | undefined,
   TServer extends Record<string, ZodType>,
   TClient extends Record<string, ZodType>,
   TShared extends Record<string, ZodType>
@@ -65,12 +67,16 @@ export interface StrictOptions<
    */
   runtimeEnvStrict: Record<
     | {
-        [TKey in keyof TClient]: TKey extends `${TPrefix}${string}`
+        [TKey in keyof TClient]: TPrefix extends undefined
+          ? never
+          : TKey extends `${TPrefix}${string}`
           ? TKey
           : never;
       }[keyof TClient]
     | {
-        [TKey in keyof TServer]: TKey extends `${TPrefix}${string}`
+        [TKey in keyof TServer]: TPrefix extends undefined
+          ? TKey
+          : TKey extends `${TPrefix}${string}`
           ? never
           : TKey;
       }[keyof TServer]
@@ -83,17 +89,18 @@ export interface StrictOptions<
 }
 
 export interface ClientOptions<
-  TPrefix extends string,
+  TPrefix extends string | undefined,
   TClient extends Record<string, ZodType>
 > {
   /**
-   * Client-side environment variables are exposed to the client by default. Set what prefix they have
+   * The prefix that client-side variables must have. This is enforced both at
+   * a type-level and at runtime.
    */
   clientPrefix: TPrefix;
 
   /**
    * Specify your client-side environment variables schema here. This way you can ensure the app isn't
-   * built with invalid env vars. To expose them to the client, prefix them with `NEXT_PUBLIC_`.
+   * built with invalid env vars.
    */
   client: Partial<{
     [TKey in keyof TClient]: TKey extends `${TPrefix}${string}`
@@ -105,7 +112,7 @@ export interface ClientOptions<
 }
 
 export interface ServerOptions<
-  TPrefix extends string,
+  TPrefix extends string | undefined,
   TServer extends Record<string, ZodType>
 > {
   /**
@@ -121,10 +128,25 @@ export interface ServerOptions<
           : never} should not prefixed with ${TPrefix}.`>
       : TServer[TKey];
   }>;
+
+  /**
+   * By default, this library will feed the environment variables directly to
+   * the Zod validator.
+   *
+   * This means that if you have an empty string for a value that is supposed
+   * to be a number (e.g. `PORT=` in a ".env" file), Zod will incorrectly flag
+   * it as a type mismatch violation. Additionally, if you have an empty string
+   * for a value that is supposed to be a string with a default value (e.g.
+   * `DOMAIN=` in an ".env" file), the default value will never be applied.
+   *
+   * In order to solve these issues, we recommend that all new projects
+   * explicitly specify this option as true.
+   */
+  emptyStringAsUndefined?: boolean;
 }
 
 export type ServerClientOptions<
-  TPrefix extends string,
+  TPrefix extends string | undefined,
   TServer extends Record<string, ZodType>,
   TClient extends Record<string, ZodType>
 > =
@@ -133,7 +155,7 @@ export type ServerClientOptions<
   | (ClientOptions<TPrefix, TClient> & Impossible<ServerOptions<never, never>>);
 
 export type EnvOptions<
-  TPrefix extends string,
+  TPrefix extends string | undefined,
   TServer extends Record<string, ZodType>,
   TClient extends Record<string, ZodType>,
   TShared extends Record<string, ZodType>
@@ -143,7 +165,7 @@ export type EnvOptions<
       ServerClientOptions<TPrefix, TServer, TClient>);
 
 export function createEnv<
-  TPrefix extends string = "",
+  TPrefix extends string | undefined,
   TServer extends Record<string, ZodType> = NonNullable<unknown>,
   TClient extends Record<string, ZodType> = NonNullable<unknown>,
   TShared extends Record<string, ZodType> = NonNullable<unknown>
@@ -157,6 +179,15 @@ export function createEnv<
   >
 > {
   const runtimeEnv = opts.runtimeEnvStrict ?? opts.runtimeEnv ?? process.env;
+
+  const emptyStringAsUndefined = opts.emptyStringAsUndefined ?? false;
+  if (emptyStringAsUndefined) {
+    for (const [key, value] of Object.entries(runtimeEnv)) {
+      if (value === "") {
+        delete runtimeEnv[key];
+      }
+    }
+  }
 
   const skip = !!opts.skipValidation;
   // eslint-disable-next-line @typescript-eslint/no-unsafe-return, @typescript-eslint/no-explicit-any
@@ -200,7 +231,12 @@ export function createEnv<
 
   const env = new Proxy(parsed.data, {
     get(target, prop) {
-      if (typeof prop !== "string" || prop === "__esModule") return undefined;
+      if (
+        typeof prop !== "string" ||
+        prop === "__esModule" ||
+        prop === "$$typeof"
+      )
+        return undefined;
       if (
         !isServer &&
         opts.clientPrefix &&